Added files via upload
"colors.li" function not working on car. 

Changing the function variable from "#Overlay" to "#truck" works great on truck overlay, just not the other way around. 

When inspecting the code online, I see the function goes to the correct place (inspector highlights the fill style link in purple when buttons are clicked), but the overlay on the car does not change. Grrrrr...